Fixed Sensor Array
I did some more fine tuning and fixed the bug that was causing the
error: Integer math that should have been floating point on line 16 of
SensorArray.cpp.

I also created an setHue() function for the RGB LED. For testing, the
LED is set to a hue that depends on the location of the line. Left =
red, middle = green, right = blue.